Changes from Lab 5
New algorithm - success! Straddles one side of the line instead of
trying to stay on it; this gives us a really easy way of telling which
way we have to turn.  Thanks to Janek for the inspiration.  Added code
to adjust motor speed with the built-in potentiometer.  To-do: play
music while winning the competition.

Also, this project will be using tabs.